% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplesIf_false.R
\name{examplesIf_false}
\alias{examplesIf_false}
\title{Examples If FALSE}
\usage{
examplesIf_false()
}
\arguments{
\item{x}{A parameter}
}
\value{
Some value
}
\description{
Examples If FALSE
}
\examples{
\dontshow{if (2 + 2 == 5) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
examplesIf_false()
\dontshow{\}) # examplesIf}
}
