% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_r6.R
\name{hello_r6}
\alias{hello_r6}
\title{Create a "conductor" tour}
\description{
blah blah blah
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Conductor-new}{\code{hello_r6$new()}}
\item \href{#method-Conductor-init}{\code{hello_r6$init()}}
\item \href{#method-Conductor-step}{\code{hello_r6$step()}}
\item \href{#method-Conductor-clone}{\code{hello_r6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Conductor-new"></a>}}
\if{latex}{\out{\hypertarget{method-Conductor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hello_r6$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Initialise \code{Conductor}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Conductor-init"></a>}}
\if{latex}{\out{\hypertarget{method-Conductor-init}{}}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hello_r6$init(session = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{A valid Shiny session. If \code{NULL} (default), the function
attempts to get the session with \code{shiny::getDefaultReactiveDomain()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialise \code{Conductor}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Conductor-step"></a>}}
\if{latex}{\out{\hypertarget{method-Conductor-step}{}}}
\subsection{Method \code{step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hello_r6$step(title = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{Title of the popover.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add a step in a \code{Conductor} tour.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Conductor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Conductor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hello_r6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
